refactor: Make Appium to a peer dependency
