+++ ci-jobs/ci.yml
@@ -1,14 +1,14 @@
 jobs:
   - job: UiAutomator2CI
     pool:
-      vmImage: macOS-10.15
+      vmImage: macOS-12
     variables:
       CI: true
       TERM: dumb
     steps:
     - task: NodeTool@0
       inputs:
-        versionSpec: 14.x
+        versionSpec: 16.x
     - script: npm install
       displayName: Install Node dependencies
     - script: npm run build
