+++ lib/commands/element.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
-import { util } from '@appium/support';
-import { PROTOCOLS, W3C_ELEMENT_KEY } from '@appium/base-driver';
+import { util } from 'appium/support';
+import { PROTOCOLS, W3C_ELEMENT_KEY } from 'appium/driver';
 
 let commands = {}, helpers = {}, extensions = {};
