+++ lib/commands/general.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import B from 'bluebird';
-import { errors } from '@appium/base-driver';
-import { fs, tempDir } from '@appium/support';
+import { errors } from 'appium/driver';
+import { fs, tempDir } from 'appium/support';
 import { APK_EXTENSION } from '../extensions';
 
 let extensions = {},
