+++ lib/commands/gestures.js
@@ -1,6 +1,6 @@
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 import _ from 'lodash';
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 
 const commands = {};
