+++ lib/commands/viewport.js
@@ -1,4 +1,4 @@
-import { imageUtil } from '@appium/support';
+import { imageUtil } from 'appium/support';
 
 
 let extensions = {}, commands = {};
