+++ lib/css-converter.js
@@ -1,6 +1,6 @@
 import { CssSelectorParser } from 'css-selector-parser';
 import { escapeRegExp } from 'lodash';
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 
 const parser = new CssSelectorParser();
 parser.registerSelectorPseudos('has');
