+++ lib/driver.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
-import { BaseDriver, DeviceSettings } from '@appium/base-driver';
+import { BaseDriver, DeviceSettings } from 'appium/driver';
 import {
   UiAutomator2Server, SERVER_PACKAGE_ID, SERVER_TEST_PACKAGE_ID
 } from './uiautomator2';
-import { fs, util, mjpeg } from '@appium/support';
+import { fs, util, mjpeg } from 'appium/support';
 import { retryInterval } from 'asyncbox';
 import B from 'bluebird';
 import commands from './commands/index';
