+++ lib/helpers.js
@@ -1,5 +1,5 @@
 import path from 'path';
-import { fs, system } from '@appium/support';
+import { fs, system } from 'appium/support';
 
 
 let helpers = {};
