+++ lib/uiautomator2.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { JWProxy, errors } from '@appium/base-driver';
+import { JWProxy, errors } from 'appium/driver';
 import { waitForCondition } from 'asyncbox';
 import {
   SERVER_APK_PATH as apkPath,
@@ -8,7 +8,7 @@ import {
 } from 'appium-uiautomator2-server';
 import {
   util, logger, tempDir, fs, timing
-} from '@appium/support';
+} from 'appium/support';
 import B from 'bluebird';
 import helpers from './helpers';
 import axios from 'axios';
