+++ package.json
@@ -40,8 +40,6 @@
     "build/lib"
   ],
   "dependencies": {
-    "@appium/base-driver": "^8.5.2",
-    "@appium/support": "^2.55.3",
     "@babel/runtime": "^7.0.0",
     "appium-adb": "^9.0.0",
     "appium-android-driver": "^5.0.15",
@@ -80,6 +78,9 @@
     "precommit-msg",
     "test"
   ],
+  "peerDependencies": {
+    "appium": "^2.0.0-beta.40"
+  },
   "devDependencies": {
     "@appium/gulp-plugins": "^7.0.0",
     "@appium/eslint-config-appium": "^6.0.0",
@@ -87,6 +88,8 @@
     "@babel/core": "^7.16.0",
     "@semantic-release/git": "^10.0.1",
     "@xmldom/xmldom": "^0.x",
+    "appium": "^2.0.0-beta.40",
+    "appium-uiautomator2-driver": "file:.",
     "android-apidemos": "^4.0.0",
     "chai": "^4.1.0",
     "chai-as-promised": "^7.1.1",
