+++ test/functional/helpers/constants.js
@@ -1,0 +1,2 @@
+export const DEFAULT_PORT = 4884;
+export const DEFAULT_HOST = '127.0.0.1';
